[Fix #994] Set `compilation-search-path` instead of advising
The advice set on `compilation-find-file` replaces the first
argument (the name of the file where the compilation error occured) with
the full path of the first file found with that name in the current
project.

When multiple files with the same name exist in a project, this may
produce the wrong path. `compilation-find-directory` handles that case
by using `compilation-directory-matcher` to track the directory of files
being compiled (by default it matches the "Entering directory" messages
produced by `make` when $(MAKE) is invoked). By replacing filenames with
full paths, the advice breaks this behaviour.

This commit replaces the advice with setting `compilation-search-path`
to the expansion of all the directories in the current project.